requirements.txt
absl-py==2.0.0
alembic==1.13.0
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
azure-core==1.29.5
azure-storage-blob==12.19.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.6
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==4.4.2
deepface==0.0.79
distro==1.8.0
filelock==3.9.0
fire==0.5.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.45.1
fsspec==2023.4.0
gast==0.5.4
gdown==4.7.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
kiwisolver==1.4.5
kombu==5.3.4
libclang==16.0.6
llvmlite==0.41.1
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
mysql==0.0.3
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.2
openai==1.3.6
openai-whisper @ git+https://github.com/openai/whisper.git@e58f28804528831904c3b6f2c0e473f346223433
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyodbc==5.0.1
pyparsing==3.1.1
pypyodbc==1.3.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.5.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
